Fix clang/win build.
TBR=jarin@chromium.org

Review URL: https://codereview.chromium.org/494973004

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@23271 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/compiler/instruction-selector.cc
@@ -382,7 +382,7 @@ void InstructionSelector::InitializeCallBuffer(Node* call, CallBuffer* buffer,
     DCHECK(deopt_node == NULL);
   }
 
-  DCHECK(input_count ==
+  DCHECK(static_cast<size_t>(input_count) ==
          (buffer->instruction_args.size() - buffer->control_count() +
           buffer->pushed_nodes.size() - buffer->frame_state_value_count()));
 }
